requirements.txt
flask==2.3.3
flask-cors==4.0.0
google-generativeai==0.3.2
speechrecognition==3.10.0
pydub==0.25.1
gunicorn==21.2.0
werkzeug==2.3.7